Issue #138 - Temp fix for overlays to display
One of the parameters is causing the overlay to not display,
'request=GetCapabilities' causes the map server to error/not send the
images needed for the overlay.

The fix now is to drop all of the query strings on the linked resources
so the overlay will display on all uploaded resources.